Resolved some float->double conversion errors to fix the tolerances for #1651
git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/laser_pipeline/trunk@26612 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ include/laser_geometry/laser_geometry.h
@@ -209,9 +209,9 @@ namespace laser_geometry
        * it is left protected so that test code can evaluate it
        * appropriately.
        */
-      const boost::numeric::ublas::matrix<double>& getUnitVectors_(float angle_min,
-                                                                   float angle_max,
-                                                                   float angle_increment,
+      const boost::numeric::ublas::matrix<double>& getUnitVectors_(double angle_min,
+                                                                   double angle_max,
+                                                                   double angle_increment,
                                                                    unsigned int length);
 
     private:
